+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Manager
   exists. Trying to create the directory now.
 * FIX: Fixed multiple execution of single schedule when multiple timed conditions match
   at the same time
+* FIX: Fixed problem in enumeration of schedules after deleting one schedule
 
 Documentation
 `````````````